Make the node builder extendable
