+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/node/HttpApiEvent.java
@@ -19,13 +19,14 @@
 package io.ballerina.flowmodelgenerator.core.model.node;
 
 import io.ballerina.flowmodelgenerator.core.model.FlowNode;
+import io.ballerina.flowmodelgenerator.core.model.NodeBuilder;
 
 /**
  * Represents the properties of a HttpApiEvent node.
  *
  * @since 1.4.0
  */
-public class HttpApiEvent extends FlowNode {
+public class HttpApiEvent extends NodeBuilder {
 
     public static final String LABEL = "HTTP API";
     public static final String EVENT_HTTP_API_METHOD = "Method";
@@ -36,13 +37,13 @@ public class HttpApiEvent extends FlowNode {
     public static final String EVENT_HTTP_API_PATH_DOC = "HTTP Path";
 
     @Override
-    public void setConstData() {
+    public void setConcreteConstData() {
         this.label = LABEL;
-        this.kind = Kind.EVENT_HTTP_API;
+        this.kind = FlowNode.Kind.EVENT_HTTP_API;
     }
 
     @Override
-    public String toSource() {
+    public String toSource(FlowNode node) {
         return null;
     }
