+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/node/If.java
@@ -20,8 +20,10 @@ package io.ballerina.flowmodelgenerator.core.model.node;
 
 import io.ballerina.compiler.syntax.tree.SyntaxKind;
 import io.ballerina.flowmodelgenerator.core.model.Branch;
-import io.ballerina.flowmodelgenerator.core.model.Property;
 import io.ballerina.flowmodelgenerator.core.model.FlowNode;
+import io.ballerina.flowmodelgenerator.core.model.NodeBuilder;
+import io.ballerina.flowmodelgenerator.core.model.Property;
+import io.ballerina.flowmodelgenerator.core.model.SourceBuilder;
 
 import java.util.List;
 import java.util.Map;
@@ -31,7 +33,7 @@ import java.util.Map;
  *
  * @since 1.4.0
  */
-public class If extends FlowNode {
+public class If extends NodeBuilder {
 
     public static final String LABEL = "If";
     public static final String DESCRIPTION = "Add conditional branch to the integration flow.";
@@ -40,26 +42,26 @@ public class If extends FlowNode {
     private static final String IF_CONDITION_DOC = "Boolean Condition";
 
     @Override
-    public void setConstData() {
+    public void setConcreteConstData() {
         this.label = LABEL;
         this.description = DESCRIPTION;
-        this.kind = Kind.IF;
+        this.kind = FlowNode.Kind.IF;
     }
 
     @Override
-    public String toSource() {
+    public String toSource(FlowNode node) {
         SourceBuilder sourceBuilder = new SourceBuilder();
-        Property condition = getProperty(Property.CONDITION_KEY);
+        Property condition = node.getProperty(Property.CONDITION_KEY);
 
         sourceBuilder
                 .keyword(SyntaxKind.IF_KEYWORD)
                 .expression(condition)
                 .openBrace();
 
-        Branch ifBranch = getBranch(IF_THEN_LABEL);
+        Branch ifBranch = node.getBranch(IF_THEN_LABEL);
         sourceBuilder.addChildren(ifBranch.children());
 
-        Branch elseBranch = getBranch(IF_ELSE_LABEL);
+        Branch elseBranch = node.getBranch(IF_ELSE_LABEL);
         if (elseBranch != null) {
             List<FlowNode> children = elseBranch.children();
             sourceBuilder
@@ -68,7 +70,7 @@ public class If extends FlowNode {
                     .keyword(SyntaxKind.ELSE_KEYWORD);
 
             // If there is only one child, and if that is an if node, generate an `else if` statement`
-            if (children.size() != 1 || children.get(0).kind() != Kind.IF) {
+            if (children.size() != 1 || children.get(0).codedata().node() != FlowNode.Kind.IF) {
                 sourceBuilder.openBrace();
             }
             sourceBuilder.addChildren(children);
@@ -80,8 +82,8 @@ public class If extends FlowNode {
 
     @Override
     public void setConcreteTemplateData() {
-        this.nodeProperties =
-                Map.of(Property.CONDITION_KEY, Property.getDefaultConditionExpression(IF_CONDITION_DOC));
+//        this.properties =
+//                Map.of(Property.CONDITION_KEY, Property.getDefaultConditionExpression(IF_CONDITION_DOC));
         this.branches = List.of(Branch.getEmptyBranch(IF_THEN_LABEL), Branch.getEmptyBranch(IF_ELSE_LABEL));
     }
 }
